+++ lib/Server.js
@@ -780,7 +780,7 @@ class Server {
     }
 
     return (
-      Server.findPort(port || this.options.port)
+      Server.getFreePort(port || this.options.port)
         // eslint-disable-next-line no-shadow
         .then((port) => {
           this.options.port = port;
@@ -846,7 +846,7 @@ class Server {
     };
   }
 
-  static findPort(port) {
+  static getFreePort(port) {
     if (port) {
       return Promise.resolve(port);
     }
