+++ test/server/Server.test.js
@@ -1,8 +1,11 @@
 'use strict';
 
 const { relative, sep } = require('path');
+const http = require('http');
 const webpack = require('webpack');
 const sockjs = require('sockjs/lib/transport');
+const portfinder = require('portfinder');
+const getFreePort = require('../../lib/Server').getFreePort;
 const Server = require('../../lib/Server');
 const config = require('../fixtures/simple-config/webpack.config');
 const port = require('../ports-map').Server;
@@ -361,6 +364,144 @@ describe('Server', () => {
       require('../../lib/Server');
 
       expect(process.env.WEBPACK_SERVE).toBe(true);
+    });
+  });
+
+  describe('getFreePort', () => {
+    let dummyServers = [];
+
+    afterEach(() => {
+      delete process.env.DEFAULT_PORT_RETRY;
+
+      return dummyServers
+        .reduce(
+          (p, server) =>
+            p.then(
+              () =>
+                new Promise((resolve) => {
+                  server.close(resolve);
+                })
+            ),
+          Promise.resolve()
+        )
+        .then(() => {
+          dummyServers = [];
+        });
+    });
+
+    function createDummyServers(n) {
+      return (Array.isArray(n) ? n : [...new Array(n)]).reduce(
+        (p, _, i) =>
+          p.then(
+            () =>
+              new Promise((resolve, reject) => {
+                const server = http.createServer();
+
+                dummyServers.push(server);
+
+                server.listen(8080 + i, () => {
+                  resolve();
+                });
+
+                server.on('error', (error) => {
+                  reject(error);
+                });
+              })
+          ),
+        Promise.resolve()
+      );
+    }
+
+    it('should returns the port when the port is specified', () => {
+      process.env.DEFAULT_PORT_RETRY = 5;
+
+      return getFreePort(8082).then((port) => {
+        expect(port).toEqual(8082);
+      });
+    });
+
+    it('should returns the port when the port is null', () => {
+      const retryCount = 2;
+
+      process.env.DEFAULT_PORT_RETRY = 2;
+
+      return createDummyServers(retryCount)
+        .then(() => getFreePort(null))
+        .then((port) => {
+          expect(port).toEqual(8080 + retryCount);
+        });
+    });
+
+    it('should returns the port when the port is undefined', () => {
+      const retryCount = 2;
+
+      process.env.DEFAULT_PORT_RETRY = 2;
+
+      return (
+        createDummyServers(retryCount)
+          // eslint-disable-next-line no-undefined
+          .then(() => getFreePort(undefined))
+          .then((port) => {
+            expect(port).toEqual(8080 + retryCount);
+          })
+      );
+    });
+
+    it('should retry finding the port for up to defaultPortRetry times (number)', () => {
+      const retryCount = 3;
+
+      process.env.DEFAULT_PORT_RETRY = retryCount;
+
+      return createDummyServers(retryCount)
+        .then(() => getFreePort())
+        .then((port) => {
+          expect(port).toEqual(8080 + retryCount);
+        });
+    });
+
+    it('should retry finding the port for up to defaultPortRetry times (string)', () => {
+      const retryCount = 3;
+
+      process.env.DEFAULT_PORT_RETRY = `${retryCount}`;
+
+      return createDummyServers(retryCount)
+        .then(() => getFreePort())
+        .then((port) => {
+          expect(port).toEqual(8080 + retryCount);
+        });
+    });
+
+    // TODO: fix me, Flaky on CI
+    it.skip('should retry finding the port when serial ports are busy', () => {
+      const busyPorts = [8080, 8081, 8082, 8083];
+
+      process.env.DEFAULT_PORT_RETRY = 3;
+
+      return createDummyServers(busyPorts)
+        .then(() => getFreePort())
+        .then((port) => {
+          expect(port).toEqual(8080 + busyPorts.length);
+        });
+    });
+
+    it("should throws the error when the port isn't found", () => {
+      expect.assertions(1);
+
+      const spy = jest
+        .spyOn(portfinder, 'getPort')
+        .mockImplementation((callback) => callback(new Error('busy')));
+
+      const retryCount = 1;
+
+      process.env.DEFAULT_PORT_RETRY = 0;
+
+      return createDummyServers(retryCount)
+        .then(() => getFreePort())
+        .catch((err) => {
+          expect(err.message).toMatchSnapshot();
+
+          spy.mockRestore();
+        });
     });
   });
 });
